_backup/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "traffic-control-openenv"
version = "0.1.0"
description = "Traffic control OpenEnv environment — deterministic intersection management benchmark"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0",
    "requests>=2.28",
    "openai>=2.0",
    "python-dotenv>=1.0",
    "huggingface_hub>=0.20",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["inference", "validate_project", "validate_submission"]
packages = ["server", "env"]